requirements.txt
python-telegram-bot==20.6
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.0
